+++ EdProject.BLL/Services/OrderService.cs
@@ -1,12 +1,14 @@
 ﻿using AutoMapper;
+using EdProject.BLL.Common.Options;
 using EdProject.BLL.Models.Base;
 using EdProject.BLL.Models.Orders;
 using EdProject.BLL.Models.Payment;
 using EdProject.BLL.Models.PrintingEditions;
 using EdProject.BLL.Services.Interfaces;
-using EdProject.DAL.DataContext;
 using EdProject.DAL.Entities;
-using EdProject.DAL.Repositories;
+using EdProject.DAL.Repositories.Interfaces;
+using Microsoft.Extensions.Options;
+using Stripe;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
@@ -16,33 +18,47 @@ namespace EdProject.BLL.Services
 {
     public class OrderService : IOrdersService
     {
-        OrderRepository _orderRepository;
-        PaymentRepository _paymentRepository;
-        EditionRepository _editionRepository;
+        IOrderRepository _orderRepository;
+        IPaymentRepository _paymentRepository;
+        IEditionRepository _editionRepository;
         IMapper _mapper;
+        private readonly StripeOptions _conectStripeOption;
 
-        public OrderService(AppDbContext appDbContext, IMapper mapper)
+        public OrderService(IMapper mapper, IOptions<StripeOptions> options,IOrderRepository orderRepository,
+                            IPaymentRepository paymentRepository,IEditionRepository editionRepository)
         {
-            _orderRepository = new OrderRepository(appDbContext);
-            _paymentRepository = new PaymentRepository(appDbContext);
-            _editionRepository = new EditionRepository(appDbContext);
+            _orderRepository = orderRepository;
+            _paymentRepository = paymentRepository;
+            _editionRepository = editionRepository;
             _mapper = mapper;
+            _conectStripeOption = options.Value;
         }
 
         public async Task CreateOrderAsync(OrderModel orderModel)
         {
-            var newOrder = _mapper.Map<OrderModel, Orders>(orderModel);
-            if(_orderRepository.OrderExist(newOrder))
+            var usersOrderList = await _orderRepository.GetOrderByUserIdAsync(orderModel.UserId);
+            if(usersOrderList.Any(order => order.StatusType == DAL.Enums.PaidStatusType.Unpaid))
             {
-                throw new CustomException(ErrorConstant.ALREADY_EXIST, HttpStatusCode.BadRequest);
+                throw new CustomException(ErrorConstant.ORDER_ERROR,HttpStatusCode.BadRequest);
             }
 
+            var newOrder = _mapper.Map<OrderModel, Orders>(orderModel);
+            newOrder.StatusType = DAL.Enums.PaidStatusType.Unpaid;
             await _orderRepository.CreateAsync(newOrder);
         }
         public async Task CreateItemInOrderAsync(OrderItemModel orderItemModel)
         {
             var item = await _editionRepository.FindByIdAsync(orderItemModel.EditionId);
             var order = await _orderRepository.FindByIdAsync(orderItemModel.OrderId);
+
+            if(order is null || order.IsRemoved)
+            {
+                throw new CustomException(ErrorConstant.INCORRECT_ORDER,HttpStatusCode.BadRequest);
+            }
+            if(item is null || item.IsRemoved)
+            {
+                throw new CustomException(ErrorConstant.ITEM_NOT_FOUND, HttpStatusCode.BadRequest);
+            }
 
             await _orderRepository.AddItemToOrderAsync(order, item);
         }
@@ -51,9 +67,11 @@ namespace EdProject.BLL.Services
             var order = await _orderRepository.FindByIdAsync(orderItemlistModel.OrderId);
             if (order is null || order.IsRemoved)
             {
-                throw new CustomException(ErrorConstant.NOTHING_EXIST, HttpStatusCode.BadRequest);
+                throw new CustomException(ErrorConstant.INCORRECT_ORDER, HttpStatusCode.BadRequest);
             }
+
             string[] editionsId = orderItemlistModel.Editions.Split(',', System.StringSplitOptions.RemoveEmptyEntries);
+
             List<Edition> editionList = new List<Edition>();
 
             foreach (var edition in editionsId)
@@ -62,7 +80,7 @@ namespace EdProject.BLL.Services
 
                 if (tempEdition is null || tempEdition.IsRemoved)
                 {
-                    throw new CustomException($"Incorrect edition Id{edition}", HttpStatusCode.BadRequest);
+                    throw new CustomException($"{ErrorConstant.CANNOT_ADD_EDITION}:{edition}", HttpStatusCode.BadRequest);
                 }
 
                 editionList.Add(tempEdition);
@@ -72,10 +90,36 @@ namespace EdProject.BLL.Services
         }
         public async Task CreatePaymentAsync(PaymentModel paymentModel)
         {
-            var newPayment = _mapper.Map<PaymentModel, Payments>(paymentModel);
-            await _paymentRepository.CreateAsync(newPayment);
             var order = await _orderRepository.FindByIdAsync(paymentModel.OrderId);
+            if (order is null || order.IsRemoved)
+            {
+                throw new CustomException(ErrorConstant.INCORRECT_ORDER, HttpStatusCode.BadRequest);
+            }
+            if (order.StatusType is DAL.Enums.PaidStatusType.Paid)
+            {
+                throw new CustomException(ErrorConstant.PAYMENT_ALREADY_PAID, HttpStatusCode.BadRequest);
+            }
 
+            StripeConfiguration.ApiKey = _conectStripeOption.SecretKey;
+            var options = new ChargeCreateOptions
+            {
+                Amount = paymentModel.Amount,
+                Currency = paymentModel.Currency.ToString().ToLower(),
+                Description = paymentModel.OrderId.ToString(),
+                Source = paymentModel.PaymentSource
+            };
+            var service = new ChargeService();
+            var charge = service.Create(options);
+            paymentModel.TransactionId = charge.Id;
+
+            if (charge.Status is not "succeeded")
+            {
+                throw new CustomException(ErrorConstant.UNSUCCESSFUL_PAYMENT, HttpStatusCode.OK);
+            }
+
+            var newPayment = _mapper.Map<PaymentModel, Payments>(paymentModel);
+            order.StatusType = DAL.Enums.PaidStatusType.Paid;
+            await _paymentRepository.CreateAsync(newPayment);
             await _orderRepository.AddPaymentToOrderAsync(order, newPayment);
         }
 
@@ -115,6 +159,7 @@ namespace EdProject.BLL.Services
                 throw new CustomException(ErrorConstant.NOTHING_EXIST, HttpStatusCode.BadRequest);
             }
             return _mapper.Map<Orders, OrderModel>(queryItem);
+                
         }
         public async Task<List<EditionModel>> GetItemsInOrderAsync(long orderId)
         {
@@ -143,7 +188,16 @@ namespace EdProject.BLL.Services
         public async Task RemoveItemFromOrderAsync(OrderItemModel orderItemModel)
         {
             var item = await _editionRepository.FindByIdAsync(orderItemModel.EditionId);
+            if(item is null)
+            {
+                throw new CustomException(ErrorConstant.ITEM_NOT_FOUND, HttpStatusCode.BadRequest); 
+            }
+
             var order = await _orderRepository.FindByIdAsync(orderItemModel.OrderId);
+            if(order is null || order.IsRemoved)
+            {
+                throw new CustomException(ErrorConstant.INCORRECT_ORDER, HttpStatusCode.BadRequest);
+            }    
 
             await _orderRepository.RemoveItemToOrderAsync(order, item);
         }
@@ -173,6 +227,16 @@ namespace EdProject.BLL.Services
             await _orderRepository.RemoveItemListFromOrderAsync(order, editionList);
         }
 
+        public async Task RemoveOrderByIdAsync(long orderId)
+        {
+            var order = await _orderRepository.FindByIdAsync(orderId);
+            if(order.Editions.Any())
+            {
+                throw new CustomException(ErrorConstant.REMOVE_ORDER_ERROR,HttpStatusCode.BadRequest);
+            }
+            order.IsRemoved = true;
+            await _orderRepository.UpdateAsync(order);
+        }
        
     }
 }
